feat : menuPage 좌측 음식 및 카테고리 컴포넌트화
